refactoring test.
