+++ integration/src/tests/build_transfer_transaction_ckb.rs
@@ -1,11 +1,10 @@
 use super::IntegrationTest;
 use crate::const_definition::{MERCURY_URI, UDT_1_HASH};
-use crate::utils::address::{
-    acp::build_acp_address, generate_rand_secp_address_pk_pair, new_identity_from_secp_address,
+use crate::utils::address::secp::{
+    generate_rand_secp_address_pk_pair, prepare_secp_address_with_ckb_capacity,
 };
-use crate::utils::instruction::{
-    issue_udt_1, prepare_account, prepare_secp_address_with_ckb_capacity, send_transaction_to_ckb,
-};
+use crate::utils::address::{acp::build_acp_address, new_identity_from_secp_address};
+use crate::utils::instruction::{issue_udt_1, prepare_account, send_transaction_to_ckb};
 use crate::utils::rpc_client::MercuryRpcClient;
 use crate::utils::signer::sign_transaction;
