+++ integration/src/tests/build_transfer_transaction_udt.rs
@@ -3,13 +3,16 @@ use crate::const_definition::{
     CHEQUE_LOCK_EPOCH, MERCURY_URI, UDT_1_HASH, UDT_1_HOLDER_ACP_ADDRESS,
     UDT_1_HOLDER_ACP_ADDRESS_PK,
 };
+use crate::utils::address::secp::{
+    generate_rand_secp_address_pk_pair, prepare_secp_address_with_ckb_capacity,
+};
 use crate::utils::address::{
     acp::build_acp_address, build_secp_address, cheque::build_cheque_address,
-    generate_rand_secp_address_pk_pair, get_udt_hash_by_owner, new_identity_from_secp_address,
+    get_udt_hash_by_owner, new_identity_from_secp_address,
 };
 use crate::utils::instruction::{
     fast_forward_epochs, issue_udt_1, issue_udt_with_cheque, prepare_account,
-    prepare_secp_address_with_ckb_capacity, send_transaction_to_ckb,
+    send_transaction_to_ckb,
 };
 use crate::utils::rpc_client::MercuryRpcClient;
 use crate::utils::signer::{sign_transaction, sign_transaction_for_cheque_of_sender};
