+++ integration/src/tests/built_in_pw_lock/build_transfer_transaction_udt.rs
@@ -2,10 +2,10 @@ use super::super::IntegrationTest;
 use crate::const_definition::{
     MERCURY_URI, UDT_1_HASH, UDT_1_HOLDER_ACP_ADDRESS, UDT_1_HOLDER_ACP_ADDRESS_PK,
 };
-use crate::utils::address::pw_lock::generate_rand_pw_address_pk_pair;
-use crate::utils::instruction::{
-    issue_udt_1, prepare_account, prepare_secp_address_with_ckb_capacity, send_transaction_to_ckb,
+use crate::utils::address::{
+    pw_lock::generate_rand_pw_address_pk_pair, secp::prepare_secp_address_with_ckb_capacity,
 };
+use crate::utils::instruction::{issue_udt_1, prepare_account, send_transaction_to_ckb};
 use crate::utils::rpc_client::MercuryRpcClient;
 use crate::utils::signer::sign_transaction;
