+++ integration/src/tests/extension_omni_lock/build_sudt_issue_transaction.rs
@@ -1,0 +1,1 @@
+
