+++ integration/src/tests/extension_omni_lock/build_transfer_transaction_ckb.rs
@@ -1,14 +1,15 @@
 use super::super::IntegrationTest;
 use crate::const_definition::{MERCURY_URI, UDT_1_HASH};
 use crate::utils::address::acp::build_acp_address;
-use crate::utils::address::generate_rand_secp_address_pk_pair;
 use crate::utils::address::omni_lock::{
     build_omni_acp_account_address, generate_omni_secp_address_pk_pair,
     prepare_omni_ethereum_address_with_capacity, prepare_omni_secp_address_with_capacity,
 };
+use crate::utils::address::secp::{
+    generate_rand_secp_address_pk_pair, prepare_secp_address_with_ckb_capacity,
+};
 use crate::utils::instruction::{
-    issue_udt_1, prepare_account, prepare_ckb_capacity, prepare_secp_address_with_ckb_capacity,
-    send_transaction_to_ckb,
+    issue_udt_1, prepare_account, prepare_ckb_capacity, send_transaction_to_ckb,
 };
 use crate::utils::rpc_client::MercuryRpcClient;
 use crate::utils::signer::sign_transaction;
