+++ integration/src/tests/extension_omni_lock/mod.rs
@@ -1,4 +1,7 @@
 mod build_adjust_account_transaction;
+mod build_dao_transaction;
+mod build_simple_transfer_transaction;
+mod build_sudt_issue_transaction;
 mod build_transfer_transaction_ckb;
 mod build_transfer_transaction_udt;
 mod get_balance;
