+++ integration/src/tests/mod.rs
@@ -1,5 +1,5 @@
 pub mod build_adjust_account_transaction;
-pub mod build_dao_related_transaction;
+pub mod build_dao_transaction;
 pub mod build_simple_transfer_transaction;
 pub mod build_sudt_issue_transaction;
 pub mod build_transfer_transaction_ckb;
