+++ integration/src/utils/address/acp.rs
@@ -19,8 +19,8 @@ pub fn build_acp_address(secp_address: &Address) -> Result<Address> {
 mod test {
     use super::*;
     use crate::const_definition::SIGHASH_TYPE_HASH;
-    use crate::utils::address::generate_rand_secp_address_pk_pair;
     use crate::utils::address::is_secp256k1;
+    use crate::utils::address::secp::generate_rand_secp_address_pk_pair;
     use std::str::FromStr;
 
     #[test]
